requirements.txt
pyodbc
fastapi
python-dotenv
jinja2
uvicorn